requirements.txt
fastapi>=0.110,<0.121
uvicorn[standard]>=0.30,<0.31
pydantic>=2.6,<2.12
python-dotenv>=1.0,<2.0
PyYAML>=6.0,<7.0
llama-index-core>=0.14,<0.15
# 0.5.6 is the first 0.5.x patch that supports llama-index-core 0.14.x,
# but it requires openai>=1.81.0 (older openai floors force pip to backtrack
# to 0.5.0, which requires llama-index-core<0.14 and breaks resolution).
llama-index-llms-openai>=0.5.6,<0.6
# 0.5.0 requires llama-index-core<0.14; 0.5.1 supports core 0.14.x.
llama-index-embeddings-openai>=0.5.1,<0.6
llama-index-vector-stores-pinecone>=0.7.1,<0.8
llama-index-vector-stores-qdrant>=0.9.1,<1.0
# Required by llama-index-vector-stores-pinecone and used by rag_v2/indexer.py.
pinecone>=7.0,<8.0
qdrant-client>=1.16,<2.0
sentence-transformers>=2.7,<2.8
openai>=1.81,<2.0
httpx>=0.27,<0.29
pytest>=8.2,<9.0
